Refactoring Java code
