--- src/main/java/no/uio/scheduler/ExecutionModeEnum.java
@@ -1,9 +1,0 @@
-package no.uio.scheduler;
-
-/**
- * Enum for the different modes of exeuction for demo and production.
- */
-public enum ExecutionModeEnum {
-    LOCAL,
-    REMOTE
-}
