+++ src/main/java/org/smolang/greenhouse/scheduler/ExecutionModeEnum.java
@@ -1,0 +1,9 @@
+package org.smolang.greenhouse.scheduler;
+
+/**
+ * Enum for the different modes of execution for demo and production.
+ */
+public enum ExecutionModeEnum {
+    LOCAL,
+    REMOTE
+}
